+++ libinet6/builder.h
@@ -100,7 +100,7 @@ int hip_build_param_notification(struct hip_common *, uint16_t, void *, size_t);
 int hip_build_param_puzzle(struct hip_common *, uint8_t, uint8_t, uint32_t,
                            uint64_t);
 int hip_build_param_r1_counter(struct hip_common *, uint64_t);
-int hip_build_param_reg_failed(struct hip_common *, uint8_t, int *, int);
+int hip_build_param_reg_failed(struct hip_common *, uint8_t, uint8_t *, int);
 int hip_build_param_reg_info(struct hip_common *, uint8_t, uint8_t, int *, int);
 int hip_build_param_reg_request(struct hip_common *, uint8_t, uint8_t[], int, int);
 int hip_build_param_rvs_hmac_contents(struct hip_common *,
